Add new `IXEvent<T>` event structures and related enums for `XInput` extension.

+++ Src/Extensions/Xcsb.Extension.XInput/Models/DeviceType.cs
@@ -1,6 +1,6 @@
 namespace Xcsb.Extension.XInput.Models;
 
-public enum DeviceType : ushort
+public enum DeviceType : byte
 {
     MasterPointer = 1,
     MasterKeyboard = 2,
